Fixed install.sh bash script
Do not override existing files

+++ install.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 mkdir -p ./web
-ln -s ./vendor/stackinstance/docker-centos/docker ./
-cp -R ./vendor/stackinstance/docker-centos/docker-compose.yml ./
-cp -R ./vendor/stackinstance/docker-centos/project.cfg ./
+
+if [ ! -f ./docker ]; then ln -s ./vendor/stackinstance/docker-centos/docker ./; fi
+if [ ! -f ./docker-compose.yml ]; then cp -R ./vendor/stackinstance/docker-centos/docker-compose.yml ./; fi
+if [ ! -f ./project.cfg ]; then cp -R ./vendor/stackinstance/docker-centos/project.cfg ./; fi
 if [ ! -f ./extra.sh ]; then cp -R ./vendor/stackinstance/docker-centos/extra.sh ./; fi
